requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Training & Evaluation
datasets>=2.14.0
evaluate>=0.4.0
wandb>=0.15.0

# NLP Utilities
tokenizers>=0.14.0
sentencepiece>=0.1.99

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
omegaconf>=2.3.0

# Kaggle
kaggle>=1.5.0
